Add files via upload
Standard Annotation: The @PreDestroy annotation is part of the Java standard library, specifically within the javax.annotation package. It is part of the Java EE (Enterprise Edition) specification and is widely supported across various frameworks, including Spring.

Lifecycle Management: It is used to manage the lifecycle of a bean, specifically to define behavior that should occur just before the bean is destroyed. This is critical for resource management, ensuring that resources are released properly.

Method Signature:

The method annotated with @PreDestroy should have a void return type.
It should not accept any parameters.
It should not throw any checked exceptions.
Multiple Methods: A single bean can have multiple lifecycle methods, but there should only be one method annotated with @PreDestroy. If multiple methods are annotated with @PreDestroy, only one will be called, and it is not guaranteed which one.

Order of Execution: The @PreDestroy method is called after the Spring container has been shut down but before the actual bean is destroyed. This is typically the last opportunity to perform any cleanup tasks.

Exception Handling: If an exception is thrown during the execution of a @PreDestroy method, the exception is logged, but it does not prevent the container from shutting down or other beans from being destroyed.

Integration with Other Frameworks: Since @PreDestroy is part of the Java EE specification, it is not limited to Spring. It can be used in other frameworks that support Java EE annotations, making it a versatile choice for managing bean lifecycles across different environments.

Alternative Approaches:

DisposableBean Interface: Spring provides the DisposableBean interface, which you can implement to achieve similar behavior. The destroy method of this interface will be called during bean destruction.
Custom Destroy Methods: You can specify a custom destroy method in the Spring configuration using the destroy-method attribute in XML configuration or using the @Bean annotation in Java-based configuration.